+++ models/demos/yolov4/README.md
@@ -4,12 +4,12 @@
 
 - Use the following command to run the yolov4 with giraffe image from weka path:
   ```
-  pytest models/experimental/yolov4/demo/demo.py
+  pytest models/demos/yolov4/demo/demo.py
   ```
 
 - Use the following command to run the yolov4 with different input image:
   ```
-  pytest  --disable-warnings --input-path=<PATH_TO_INPUT_IMAGE> models/experimental/yolov4/demo/demo.py
+  pytest  --disable-warnings --input-path=<PATH_TO_INPUT_IMAGE> models/demos/yolov4/demo/demo.py
   ```
 
 Once you run the command, The output file named `ttnn_prediction_demo.jpg` will be generated.
