+++ models/demos/yolov4/demo/giraffe_320.jpg
