+++ models/demos/yolov4/reference/downsample2.py
@@ -5,7 +5,7 @@
 
 import torch
 import torch.nn as nn
-from models.experimental.yolov4.reference.resblock import ResBlock
+from models.demos.yolov4.reference.resblock import ResBlock
 
 
 class Mish(torch.nn.Module):
@@ -17,28 +17,26 @@ class Mish(torch.nn.Module):
         return x
 
 
-class DownSample5(nn.Module):
+class DownSample2(nn.Module):
     def __init__(self):
         super().__init__()
-        self.c1 = nn.Conv2d(512, 1024, 3, 2, 1, bias=False)
-        self.b1 = nn.BatchNorm2d(1024)
+        self.c1 = nn.Conv2d(64, 128, 3, 2, 1, bias=False)
+        self.b1 = nn.BatchNorm2d(128)
         self.mish = Mish()
 
-        self.c2 = nn.Conv2d(1024, 512, 1, 1, bias=False)
-        self.b2 = nn.BatchNorm2d(512)
+        self.c2 = nn.Conv2d(128, 64, 1, 1, 0, bias=False)
+        self.b2 = nn.BatchNorm2d(64)
 
-        self.c3 = nn.Conv2d(1024, 512, 1, 1, bias=False)
-        self.b3 = nn.BatchNorm2d(512)
+        self.c3 = nn.Conv2d(128, 64, 1, 1, 0, bias=False)
+        self.b3 = nn.BatchNorm2d(64)
 
-        self.res = ResBlock(512, 4)
+        self.res = ResBlock(ch=64, nblocks=2)
 
-        self.c4 = nn.Conv2d(512, 512, 1, 1, bias=False)
-        self.b4 = nn.BatchNorm2d(512)
-        self.mish = Mish()
+        self.c4 = nn.Conv2d(64, 64, 1, 1, 0, bias=False)
+        self.b4 = nn.BatchNorm2d(64)
 
-        self.c5 = nn.Conv2d(1024, 1024, 1, 1, bias=False)
-        self.b5 = nn.BatchNorm2d(1024)
-        self.mish = Mish()
+        self.c5 = nn.Conv2d(128, 128, 1, 1, 0, bias=False)
+        self.b5 = nn.BatchNorm2d(128)
 
     def forward(self, input: torch.Tensor):
         x1 = self.c1(input)
@@ -53,10 +51,9 @@ class DownSample5(nn.Module):
         x3_b = self.b3(x3)
         x3_m = self.mish(x3_b)
 
-        # resblock
-        r = self.res(x3_m)
+        r1 = self.res(x3_m)
 
-        x4 = self.c4(r)
+        x4 = self.c4(r1)
         x4_b = self.b4(x4)
         x4_m = self.mish(x4_b)
 
@@ -65,5 +62,4 @@ class DownSample5(nn.Module):
         x5 = self.c5(x4_m)
         x5_b = self.b5(x5)
         x5_m = self.mish(x5_b)
-
         return x5_m
