+++ models/demos/yolov4/ttnn/neck.py
@@ -4,7 +4,7 @@
 
 import torch
 import ttnn
-from models.experimental.yolov4.ttnn.common import Conv
+from models.demos.yolov4.ttnn.common import Conv
 from tt_lib.fallback_ops import fallback_ops
