+++ models/demos/yolov4/ttnn/yolov4.py
@@ -14,13 +14,13 @@ ttnn.enable_detailed_buffer_report = True
 ttnn.enable_detailed_tensor_report = True
 ttnn.enable_comparison_mode = False
 
-from models.experimental.yolov4.ttnn.downsample1 import Down1
-from models.experimental.yolov4.ttnn.downsample2 import Down2
-from models.experimental.yolov4.ttnn.downsample3 import Down3
-from models.experimental.yolov4.ttnn.downsample4 import Down4
-from models.experimental.yolov4.ttnn.downsample5 import Down5
-from models.experimental.yolov4.ttnn.neck import TtNeck
-from models.experimental.yolov4.ttnn.head import TtHead
+from models.demos.yolov4.ttnn.downsample1 import Down1
+from models.demos.yolov4.ttnn.downsample2 import Down2
+from models.demos.yolov4.ttnn.downsample3 import Down3
+from models.demos.yolov4.ttnn.downsample4 import Down4
+from models.demos.yolov4.ttnn.downsample5 import Down5
+from models.demos.yolov4.ttnn.neck import TtNeck
+from models.demos.yolov4.ttnn.head import TtHead
 
 
 class TtYOLOv4:
